Fix recurring docs example
Also, I notice that the example doesn't work if there's a job failure (as run_at would change). I think to do this properly, you'd want to put in the time range the users should be checked in the job args. Or, we could add a 'created_at' timestamptz to the que_jobs table and use that to figure out the range of time -- as that wouldn't change if there's a failure.