use now() instead of 'now'
The argument 'now' is evaluated when the statement is prepared, not when
it is executed. So the timestamp in the insert_job query will be stale
after the first execution.

Switching it to now() ensures that it is always evaluated when the query
is executed.

See https://github.com/bgentry/que-go/issues/1 for more details.