Document Que's `lock_job` mechanism
I had a bit of a hard time understanding exactly how this locking
function worked when I first examined it. This pull adds a little bit of
a explanatory documentation in an attempt to make it a little easier to
digest.

This may be redundant to the SQL elite among us, but could potentially
be helpful to people who are more casually browsing the repository.