Tweak the job lock query very slightly.
In the previous version of the job lock query, the job.locked filter in the second half of the recursive UNION
ALL statement functioned to halt the CTE as soon as a locked job was found, which rendered the LIMIT clause in
the main SELECT statement meaningless, since the rCTE could only ever return one job (or no jobs) anyway. Some
time ago, I modified it in the 1.0 branch to only stop recursing when no more jobs were available at all, and
to actually make use of the LIMIT clause. This was done to support batch polling, allowing Que to lock exactly
N jobs by passing a LIMIT N.

Long story short, this rephrasing of the query is equivalent, but more understandable, better matches the code
comment just added via PR, and can be easily adapted to lock jobs in batches by anyone copy-pasting the query.