Docs index via README
I'm looking at using Que in a web app I'm working on, but I see this project doesn't have a wiki and there's no index for the docs. This makes it a bit difficult for me to find what I'm looking for, so I started a basic index. It's a list of links to each doc with a sublist of with any subheadings for that doc.

I'd recommend adding short descriptions (a sentence or two perhaps) to the index to make it a bit clearer as to what is covered in each topic.