Handle errors with blank message when setting last_error.
If an error is raised with a blank message, as (for example) sometimes
`OpenSSL::Cipher::CipherError` can be, the `last_error` that is saved
to the DB for that job has a confusing first blank line. For older
versions of Que (before 3feebc56738a3b114472386a4c68e3ac1b541a4c) with
no stacktrace, all you get is an empty string.