Fix internal document links
Attempt to fix links to point to other parts of this document, not sure I got them all.

This document was originally just an index that pointed to other documents that made up the docs. However, after the documents were merged, it looks like the internal hyperlinks were not fixed.

Also fixed at least one link that wasn't the correct link.