Add more context to error message when config files fail to load.
Currently when a configuration/app file is required, if it fails for any reason, then a simple "Could not load file" message is given.

This might be because the required file is missing, or it could be because it contains errors.

This commit adds more context to the reason for the failure, helping debug cases where the required file contains an error.