Drop support for Ruby < 2.7, and hence Rails < 6.0
In adding support for Ruby 3, we need to split keyword arguments of out the args. [It's too difficult to support three different keyword argument behaviours (Ruby 2.6, 2.7, 3)](https://eregon.me/blog/2019/11/10/the-delegation-challenge-of-ruby27.html). Since Ruby 2.6 is [coming up for End-of-Life in just three months (on 2022-03-31)](https://www.ruby-lang.org/en/downloads/branches/), we reckon it was fine to drop support for it.

Co-Authored-By: Brendan Weibrecht <brendan@weibrecht.net.au>